backend/requirements.txt
Flask-CORS
Flask
python-dotenv
openai==0.28.0
